requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.7
pandas==2.1.4
openpyxl==3.1.2
sqlalchemy==2.0.23